♻️ [Refactoring]: DIコンテナのimportパスを新構造に更新
- bindings.tsとfactories.tsのimportパスを統合インターフェースからの参照に変更
- 新しいフォルダ構造に対応したクリーンな依存関係に更新
